moved
moved images in subscription to src/assets/images
moved components to src/components and created a subscription folder for subscription page components
