+++ src/router/index.tsx
@@ -2,45 +2,42 @@ import { useRoutes } from "react-router-dom";
 
 import Dashboard from "../Pages/Dashboard";
 import ForgotPassword from "../Pages/Auth/ForgotPassword";
-import PrivateRoute from './PrivateRoute';
-
-
+import PrivateRoute from "./PrivateRoute";
+import SubscriptionPage from "../Pages/subscription";
 
 function Router() {
   const routes = [
-   
     {
       path: "/",
       element: <Dashboard />,
-     
     },
-    {path: "/forgot-password", element: <ForgotPassword/>},
+    { path: "/forgot-password", element: <ForgotPassword /> },
     // {path: "/scan", element: <QrReader/>},
 
-  
     {
       // path: "/dashboard",
       element: <PrivateRoute />,
       children: [
-        
-     
         // {
         //   path: "/dashboard",
         //   element: <Dashboard />,
-        // },  
-
+        // },
       ],
     },
 
-
-
-    
-
-   
+    {
+      path: "/subscription",
+      element: <PrivateRoute />,
+      children: [
+        {
+          path: "/subscription",
+          element: <SubscriptionPage />,
+        },
+      ],
+    },
   ];
 
   return useRoutes(routes);
 }
 
 export default Router;
-
